[REFACTOR] 파일위치 및 분리 작업
