+++ FinalTodo/FinalTodo/Helper/Extensions.swift
@@ -28,3 +28,36 @@ extension UIView {
         layer.mask = mask
     }
 }
+
+// Date 객체의 확장을 통해 특정 날짜의 시작 시간을 가져오는 기능
+extension Date {
+    // 해당 날짜의 시작 시간을 반환
+    var startOfDay: Date {
+        return Calendar.current.startOfDay(for: self)
+    }
+}
+
+extension UIView {
+    func shake() {
+        self.transform = CGAffineTransform(translationX: 20, y: 0)
+        UIView.animate(withDuration: 0.4, delay: 0, usingSpringWithDamping: 0.2, initialSpringVelocity: 1, options: .curveEaseInOut, animations: {
+            self.transform = CGAffineTransform.identity
+        }, completion: nil)
+    }
+}
+
+extension UINavigationController {
+    func popViewController(animated: Bool, completion:@escaping (()->())) {
+        CATransaction.setCompletionBlock(completion)
+        CATransaction.begin()
+        _ = self.popViewController(animated: animated)
+        CATransaction.commit()
+    }
+    
+    func pushViewController(_ viewController: UIViewController, animated: Bool, completion:@escaping (()->())) {
+        CATransaction.setCompletionBlock(completion)
+        CATransaction.begin()
+        self.pushViewController(viewController, animated: animated)
+        CATransaction.commit()
+    }
+}
