+++ FinalTodo/FinalTodo/Model/User.swift
@@ -7,7 +7,7 @@
 
 import UIKit
 
-struct Users: Codable {
+struct User: Codable {
     let id: String
     let nickName: String
     let folders: [Folders]
